requirements.txt
openai>=1.2.4
python-dotenv
flask
line-bot-sdk
